Split opcodes logic into codes, fns, and utils files
