+++ packages/vm/lib/evm/opcodes/index.ts
@@ -1,0 +1,3 @@
+export * from './codes'
+export * from './functions'
+export * from './util'
